refactor(Gallery): Move styles
